Split controllers to make them smaller
